Linked API with validator class
