+++ cg/apps/demultiplex/sample_sheet/override_cycles_validator.py
@@ -1,0 +1,145 @@
+import logging
+import re
+
+from cg.exc import OverrideCyclesError
+
+LOG = logging.getLogger(__name__)
+FORWARD_INDEX_CYCLE_PATTERN: str = r"I(\d+)N(\d+)"
+REVERSE_INDEX_CYCLE_PATTERN: str = r"N(\d+)I(\d+)"
+
+
+class OverrideCyclesValidator:
+    """Class for validating the override cycles value of a sample in a sample sheet."""
+
+    def __init__(
+        self,
+        run_read1_cycles: int,
+        run_read2_cycles: int,
+        run_index1_cycles: int,
+        run_index2_cycles: int | None,
+        is_reverse_complement: bool,
+    ):
+        self.sample: dict[str, str] | None = None
+        self.sample_cycles: list[str] | None = None
+        self.sample_id: str | None = None
+        self.run_read1_cycles: int = run_read1_cycles
+        self.run_read2_cycles: int = run_read2_cycles
+        self.run_index1_cycles: int = run_index1_cycles
+        self.run_index2_cycles: int | None = run_index2_cycles
+        self.is_reverse_complement: bool = is_reverse_complement
+
+    @staticmethod
+    def is_index_cycle_value_following_pattern(
+        pattern: str, index_cycle: str, run_cycles: int, index_sequence: str
+    ) -> bool:
+        """
+        Returns whether an index cycle string is following a valid cycle regex pattern and has
+        consistent values. Valid patterns are 'I(\d+)N(\d+)' and 'N(\d+)I(\d+)'. Having consistent
+        values means that the sum of the number of index characters (I) and the number of ignored
+        characters (N) specified in the index cycle string is equal to the number of run cycles
+        and the length of the index sequence is equal to the number of index characters (I).
+        """
+        match = re.match(pattern, index_cycle)
+        if match:
+            if pattern == FORWARD_INDEX_CYCLE_PATTERN:
+                index_chars, ignored_chars = map(int, match.groups())
+            elif pattern == REVERSE_INDEX_CYCLE_PATTERN:
+                ignored_chars, index_chars = map(int, match.groups())
+            else:
+                LOG.warning(f"Pattern {pattern} is not a valid index cycle pattern")
+                return False
+            if index_chars + ignored_chars == run_cycles and len(index_sequence) == index_chars:
+                return True
+        return False
+
+    def _validate_reads_cycles(self) -> None:
+        """
+        Determine if the sample read cycles are valid, i.e. if the sample read cycle values are
+        equal to the respective run read cycles.
+        Raises:
+            OverrideCyclesError if the reads cycles are not valid.
+        """
+        read1_cycle: str = self.sample_cycles[0]
+        read2_cycle: str = self.sample_cycles[-1]
+        if (
+            read1_cycle == f"Y{self.run_read1_cycles}"
+            and read2_cycle == f"Y{self.run_read2_cycles}"
+        ):
+            return
+        message: str = f"Incorrect read cycles {self.sample_cycles} for sample {self.sample_id}"
+        LOG.error(message)
+        raise OverrideCyclesError(message)
+
+    def _validate_index1_cycles(self) -> None:
+        """
+        Determine if the sample index 1 cycle is valid, i.e., if the number of index characters in
+        the override cycles coincides with the length of the index sequence and if the number of
+        ignored characters in the override cycles matches the difference between the length of the
+        index sequence and the number of run index1 cycles.
+        Raises:
+            OverrideCyclesError if the index 1 cycle is not valid.
+        """
+        index1_cycle: str = self.sample_cycles[1]
+        if (
+            self.run_index1_cycles == len(self.sample["Index"])
+            and index1_cycle == f"I{self.run_index1_cycles}"
+        ):
+            return
+        if self.is_index_cycle_value_following_pattern(
+            pattern=FORWARD_INDEX_CYCLE_PATTERN,
+            index_cycle=index1_cycle,
+            run_cycles=self.run_index1_cycles,
+            index_sequence=self.sample["Index"],
+        ):
+            return
+        message: str = f"Incorrect index1 cycle {index1_cycle} for sample {self.sample_id}"
+        LOG.error(message)
+        raise OverrideCyclesError(message)
+
+    def _validate_index2_cycles(self) -> None:
+        """
+        Determine if the index 2 cycle is valid, i.e., if the number of ignored and index characters
+        correspond to the length of the sample index2 sequence and the number of run index2 cycles,
+        or if the index cycles should be None.
+        Raises:
+            OverrideCyclesError if the index 2 cycle is not valid.
+        """
+        if not self.run_index2_cycles and len(self.sample_cycles) == 3:
+            return
+        index2_cycle: str = self.sample_cycles[2]
+        if not self.sample["Index2"] and index2_cycle == f"N{self.run_index2_cycles}":
+            return
+        if (
+            self.run_index2_cycles == len(self.sample["Index2"])
+            and index2_cycle == f"I{self.run_index2_cycles}"
+        ):
+            return
+        if self.is_reverse_complement and self.is_index_cycle_value_following_pattern(
+            pattern=REVERSE_INDEX_CYCLE_PATTERN,
+            index_cycle=index2_cycle,
+            run_cycles=self.run_index2_cycles,
+            index_sequence=self.sample["Index2"],
+        ):
+            return
+        if not self.is_reverse_complement and self.is_index_cycle_value_following_pattern(
+            pattern=FORWARD_INDEX_CYCLE_PATTERN,
+            index_cycle=index2_cycle,
+            run_cycles=self.run_index2_cycles,
+            index_sequence=self.sample["Index2"],
+        ):
+            return
+        message: str = f"Incorrect index2 cycle {index2_cycle} for sample {self.sample_id}"
+        LOG.error(message)
+        raise OverrideCyclesError(message)
+
+    def validate_sample(
+        self,
+        sample: dict[str, str],
+    ) -> None:
+        """Determine if the override cycles are valid for a given sample."""
+        self.sample = sample
+        self.sample_cycles = sample["OverrideCycles"].split(";")
+        self.sample_id = sample["Sample_ID"]
+        self._validate_reads_cycles()
+        self._validate_index1_cycles()
+        self._validate_index2_cycles()
