+++ cg/apps/demultiplex/sample_sheet/read_sample_sheet.py
@@ -38,24 +38,8 @@ def validate_samples_unique_per_lane(samples: list[FlowCellSample]) -> None:
         validate_samples_are_unique(samples=lane_samples)
 
 
-def get_sample_sheet_from_file(infile: Path) -> SampleSheet:
-    """Parse and validate a sample sheet from file."""
-    sample_sheet_content: list[list[str]] = ReadFile.get_content_from_file(
-        file_format=FileFormat.CSV, file_path=infile
-    )
-    sample_type: Type[FlowCellSample] = get_sample_type(infile)
-
-    return get_validated_sample_sheet(
-        sample_sheet_content=sample_sheet_content,
-        sample_type=sample_type,
-    )
-
-
-def get_sample_type(sample_sheet_path: Path) -> Type[FlowCellSample]:
-    """Returns the sample type based on the header of the given sample sheet."""
-    sample_sheet_content: list[list[str]] = ReadFile.get_content_from_file(
-        file_format=FileFormat.CSV, file_path=sample_sheet_path
-    )
+def get_sample_type(sample_sheet_content: list[list[str]]) -> Type[FlowCellSample]:
+    """Returns the sample type identified from the sample sheet content."""
     for row in sample_sheet_content:
         if not row:
             continue
@@ -65,19 +49,9 @@ def get_sample_type(sample_sheet_path: Path) -> Type[FlowCellSample]:
         if SampleSheetBcl2FastqSections.Data.HEADER in row[0]:
             LOG.info("Sample sheet was generated for BCL2FASTQ")
             return FlowCellSampleBcl2Fastq
-    raise SampleSheetError("Could not determine sample sheet type")
-
-
-def get_validated_sample_sheet(
-    sample_sheet_content: list[list[str]],
-    sample_type: Type[FlowCellSample],
-) -> SampleSheet:
-    """Return a validated sample sheet object."""
-    raw_samples: list[dict[str, str]] = get_raw_samples(sample_sheet_content=sample_sheet_content)
-    adapter = TypeAdapter(list[sample_type])
-    samples = adapter.validate_python(raw_samples)
-    validate_samples_unique_per_lane(samples=samples)
-    return SampleSheet(samples=samples)
+    message: str = "Could not determine sample sheet type"
+    LOG.error(message)
+    raise SampleSheetError(message)
 
 
 def get_raw_samples(sample_sheet_content: list[list[str]]) -> list[dict[str, str]]:
@@ -120,3 +94,30 @@ def get_samples_by_lane(
             sample_by_lane[sample.lane] = []
         sample_by_lane[sample.lane].append(sample)
     return sample_by_lane
+
+
+def get_flow_cell_samples_from_content(
+    sample_sheet_content: list[list[str]],
+) -> list[FlowCellSample]:
+    """Return the samples in a sample sheet as a list of FlowCellSample objects."""
+    sample_type: Type[FlowCellSample] = get_sample_type(sample_sheet_content)
+    raw_samples: list[dict[str, str]] = get_raw_samples(sample_sheet_content=sample_sheet_content)
+    adapter = TypeAdapter(list[sample_type])
+    return adapter.validate_python(raw_samples)
+
+
+def get_validated_sample_sheet(
+    sample_sheet_content: list[list[str]],
+) -> SampleSheet:
+    """Return a validated sample sheet object."""
+    samples: list[FlowCellSample] = get_flow_cell_samples_from_content(sample_sheet_content)
+    validate_samples_unique_per_lane(samples=samples)
+    return SampleSheet(samples=samples)
+
+
+def get_sample_sheet_from_file(infile: Path) -> SampleSheet:
+    """Parse and validate a sample sheet from file."""
+    sample_sheet_content: list[list[str]] = ReadFile.get_content_from_file(
+        file_format=FileFormat.CSV, file_path=infile
+    )
+    return get_validated_sample_sheet(sample_sheet_content)
